fix: Change NetCDF encoding from float16 to float32
NetCDF4 library doesn't support float16 data type. Changed encoding
to float32 which is supported and still provides space savings vs float64.

Fixes TypeError: Illegal primitive data type error when saving spectrograms.

+++ examples/aplose_audio_processor/generator.py
@@ -277,9 +277,9 @@ class AploseAudioProcessor:
             }
         )
 
-        # Save NetCDF with float16 encoding
+        # Save NetCDF with float32 encoding
         netcdf_path = str(Path(wav_path).with_suffix('.nc'))
-        encoding = {'spectrogram': {'dtype': 'float16'}}
+        encoding = {'spectrogram': {'dtype': 'float32'}}
         ds.to_netcdf(netcdf_path, encoding=encoding)
 
         return netcdf_path
@@ -346,9 +346,9 @@ class AploseAudioProcessor:
             }
         )
 
-        # Save NetCDF with float16 encoding for all spectrogram variables
+        # Save NetCDF with float32 encoding for all spectrogram variables
         netcdf_path = str(Path(wav_path).with_suffix('.nc'))
-        encoding = {var_name: {'dtype': 'float16'} for var_name in data_vars.keys()}
+        encoding = {var_name: {'dtype': 'float32'} for var_name in data_vars.keys()}
         ds.to_netcdf(netcdf_path, encoding=encoding)
 
         return netcdf_path
